pyproject.toml
[project]
name = "finance-planner-cli"
version = "0.1.0"
description = "Developer CLI for Finance Planner and MCP_Auth services"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dateutil>=2.8.2",
    "shellingham>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
finance-cli = "cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["cli"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
